jsonrpc: convert get-accumulator-consistency-proof test to forge
